+++ Numismatics.CORE/Domain/Models/Banknote.cs
@@ -9,9 +9,9 @@ namespace Numismatics.CORE.Domain.Models
 {
     public class Banknote
     {
-        public int Id { get; set; }
-        public int CountryId { get; set; }
-        public int CurrencyId { get; set; }
+        public long Id { get; set; }
+        public long CountryId { get; set; }
+        public long CurrencyId { get; set; }
         public double Value { get; set; }
         public bool HundertPart { get; set; }
         public Date IssueDate { get; set; }
@@ -22,7 +22,7 @@ namespace Numismatics.CORE.Domain.Models
         public Dictionary<string, MoneyQuality> Banknotes { get; set; }
         public Banknote() { }
 
-        public Banknote(int id, int countryId, int currencyId, double value, bool hundertPart, Date issueDate, string obversePicture, string reversePicture, string description, string city, Dictionary<string, MoneyQuality> banknotes)
+        public Banknote(long id, long countryId, long currencyId, double value, bool hundertPart, Date issueDate, string obversePicture, string reversePicture, string description, string city, Dictionary<string, MoneyQuality> banknotes)
         {
             Id = id;
             CountryId = countryId;
